Agrego y cambio en controladores, modelos y servicios de Productos y Usuarios
Agrego validaciones en productoController.js (valido que req.body no es nulo)
Agrego validaciones en authController.js (valido que req.body no es nulo)
Cambio el método register de usuarioService.js

Agrego validaciones en métodos de productoModel.js (valida campos de JSON producto en saveProduc), (valida valor id de getProductById y deleteProduct)

Agrego validaciones en métodos de usuarioModel.js (valida campos de JSON usuario)

Agrego método loginUser en usuarioModel.js
(Combina validateEmail y validatePassword)

+++ models/productoModel.js
@@ -8,12 +8,12 @@ const productsCollection = collection(db, 'productos');
 
 export async function getProductById(id) {
     try {
-        const productDoc = await getDoc(doc(productsCollection, id));
-        
-        if(isNaN(Number(id)) || Number(id) < 0) {
+        if(id === "" || isNaN(Number(id)) || Number(id) < 0) {
             throw new ConflictError('Error obteniendo producto: el id debe ser un valor númerico válido');
         }
         
+        const productDoc = await getDoc(doc(productsCollection, id));
+
         if (productDoc.exists()) {
             return productDoc.data();
         } else {
@@ -45,16 +45,16 @@ export async function getAllProducts() {
 
 export async function saveProduct(product) {
     try {
-        if(product.id) {
-            const productDoc = await getDoc(doc(productsCollection, product.id));
+        if(!product.id || !product.nombre || !product.precio) {
+            throw new ConflictError("Error creando producto: debe ingresar el ID, nombre y precio del producto");
+        }
+
+        const productDoc = await getDoc(doc(productsCollection, product.id));
         
-            if (productDoc.exists()) {
-                throw new ConflictError('Error creando producto: ya existe un producto con ese id');
-            } else {
-                await setDoc(doc(productsCollection, product.id), product);
-            }
+        if (productDoc.exists()) {
+            throw new ConflictError('Error creando producto: El ID del producto ya existe o es inválido');
         } else {
-            throw new ConflictError('Error creando producto: producto con id nulo');
+            await setDoc(doc(productsCollection, product.id), product);
         }
     } catch (error) {
         throw error;
@@ -64,6 +64,11 @@ export async function saveProduct(product) {
 export async function deleteProduct(id) {
     try {
         const productDoc = await getDoc(doc(productsCollection, id));
+        
+        if(id === "" || isNaN(Number(id)) || Number(id) < 0) {
+            throw new ConflictError('Error obteniendo producto: el ID debe ser un valor númerico válido');
+        }
+
         if (productDoc.exists()) {
             await deleteDoc(doc(productsCollection, id));
         } else {
